iot_app/requirements.txt
# ===============================
# Core Communication
# ===============================
paho-mqtt==1.6.1
mysql-connector-python==8.3.0

# ===============================
# Logging
# ===============================
loguru==0.7.2

# ===============================
# GUI Options
# ===============================
PyQt5==5.15.9

# ===============================
# Data Handling & Visualization
# ===============================
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2

# ===============================
# Optional Dev Tools
# ===============================
# black==24.4.2
# isort==5.13.2
# pytest==8.2.1
